# Faster model with dynamic launch configuration by @maleadt