--- a/src/diagnostics.jl
+++ b/src/diagnostics.jl
@@ -46,7 +46,7 @@ struct VelocityDivergenceChecker <: Diagnostic
 end
 
 function velocity_div!(grid::RegularCartesianGrid, u, v, w, div)
-    @loop for k in (1:grid.Nz; blockIdx().z)
+    @loop for k in (1:grid.Nz; (blockIdx().z - 1) * blockDim().z + threadIdx().z)
         @loop for j in (1:grid.Ny; (blockIdx().y - 1) * blockDim().y + threadIdx().y)
             @loop for i in (1:grid.Nx; (blockIdx().x - 1) * blockDim().x + threadIdx().x)
                 @inbounds div[i, j, k] = div_f2c(grid, u, v, w, i, j, k)