This PR integrates Tim's changes from the GPU hackathon. By using dynamic launch configurations and splitting up the interior source term calculation kernel (+ https://github.com/JuliaGPU/CUDAnative.jl/pull/417) we get a pretty sweet ~25% overall speedup (so the most expensive kernel which was a bottleneck is sped up by ~40%).

Needs https://github.com/vchuravy/GPUifyLoops.jl/pull/90 to work for now.

cc @maleadt Sorry it took so long for me to merge these improvements.

Actually resolves #64 

```
──────────────────────────────────────────────────────────────────────────────────────
 Oceananigans.jl static ocean bench...        Time                   Allocations      
                                      ──────────────────────   ───────────────────────
           Tot / % measured:                409s / 28.8%           15.0GiB / 0.50%    

 Section                      ncalls     time   %tot     avg     alloc   %tot      avg
 ─────────────────────────────────────────────────────────────────────────────────────
 256×256×256 (CPU, Float64)       10    56.0s  47.7%   5.60s    292KiB  0.37%  29.2KiB
 256×256×256 (CPU, Float32)       10    47.2s  40.2%   4.72s    227KiB  0.29%  22.7KiB
 128×128×128 (CPU, Float32)       10    5.91s  5.03%   591ms    227KiB  0.29%  22.7KiB
 128×128×128 (CPU, Float64)       10    5.87s  5.00%   587ms    292KiB  0.37%  29.2KiB
  64× 64× 64 (CPU, Float64)       10    803ms  0.68%  80.3ms    292KiB  0.37%  29.2KiB
  64× 64× 64 (CPU, Float32)       10    724ms  0.62%  72.4ms    227KiB  0.29%  22.7KiB
 256×256×256 (GPU, Float64)       10    309ms  0.26%  30.9ms   9.83MiB  12.9%  0.98MiB
 256×256×256 (GPU, Float32)       10    239ms  0.20%  23.9ms   8.70MiB  11.4%   891KiB
  32× 32× 32 (CPU, Float64)       10   86.6ms  0.07%  8.66ms    292KiB  0.37%  29.2KiB
  32× 32× 32 (CPU, Float32)       10   61.0ms  0.05%  6.10ms    227KiB  0.29%  22.7KiB
  32× 32× 32 (GPU, Float64)       10   50.7ms  0.04%  5.07ms   9.83MiB  12.9%  0.98MiB
  64× 64× 64 (GPU, Float64)       10   49.7ms  0.04%  4.97ms   9.83MiB  12.9%  0.98MiB
 128×128×128 (GPU, Float64)       10   46.9ms  0.04%  4.69ms   9.83MiB  12.9%  0.98MiB
  32× 32× 32 (GPU, Float32)       10   44.8ms  0.04%  4.48ms   8.70MiB  11.4%   891KiB
 128×128×128 (GPU, Float32)       10   40.6ms  0.03%  4.06ms   8.70MiB  11.4%   891KiB
  64× 64× 64 (GPU, Float32)       10   32.6ms  0.03%  3.26ms   8.70MiB  11.4%   891KiB
 ─────────────────────────────────────────────────────────────────────────────────────

CPU Float64 -> Float32 speedup:
 32× 32× 32: 1.419
 64× 64× 64: 1.110
128×128×128: 0.993
256×256×256: 1.186

GPU Float64 -> Float32 speedup:
 32× 32× 32: 1.132
 64× 64× 64: 1.524
128×128×128: 1.155
256×256×256: 1.294

CPU -> GPU speedup:
 32× 32× 32 (Float32): 1.363
 32× 32× 32 (Float64): 1.710
 64× 64× 64 (Float32): 22.178
 64× 64× 64 (Float64): 16.151
128×128×128 (Float32): 145.438
128×128×128 (Float64): 125.088
256×256×256 (Float32): 197.674
256×256×256 (Float64): 181.197
```